+++ headers/heartbeat.h
@@ -1,6 +1,6 @@
 #ifndef HEARTBEAT_H
 #define HEARTBEAT_H
-VAR uint16_t Heartbeat_Delay;
+VAR uint32_t Heartbeat_Delay;
 VAR bool Heartbeat_Enabled;
 VAR const char* Heartbeat_URL;
 char Heartbeat_Secret[17];
